Move dot_file to its own module
